backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
groq
cerebras-cloud-sdk
chromadb
httpx
python-multipart
deepgram-sdk
